refactor(ui/result): opt code structure
