--- packages/varlet-ui/src/result-item/example/index.vue
@@ -1,12 +1,0 @@
-<script setup>
-import VarResultItem from '../index'
-import { watchLang, AppType } from '@varlet/cli/client'
-import { use, pack } from './locale'
-
-watchLang(use)
-</script>
-
-<template>
-  <app-type>Mobile phone example code</app-type>
-  <var-result-item />
-</template>
