--- packages/varlet-ui/src/result-item/example/locale/en-US.ts
@@ -1,3 +1,0 @@
-export default {
-  // Example locale
-}
