--- packages/varlet-ui/src/result-item/example/locale/zh-CN.ts
@@ -1,3 +1,0 @@
-export default {
-  // Example locale
-}
