--- packages/varlet-ui/src/result-item/props.ts
@@ -1,12 +1,0 @@
-export const props = {
-  imageSize: {
-    type: [String, Number],
-    default: 80,
-  },
-  status: {
-    type: String,
-  },
-  duration: {
-    type: [Number, String],
-  },
-}
