--- packages/varlet-ui/src/result-item/resultItem.less
@@ -1,223 +1,0 @@
-:root {
-  --result-item-info-color: var(--color-info);
-  --result-item-info-border-color: #00afef4c;
-  --result-item-success-color: var(--color-success);
-  --result-item-success-border-color: #00c48f4c;
-  --result-item-error-color: var(--color-danger);
-  --result-item-error-border-color: #f443364d;
-  --result-item-warning-color: var(--color-warning);
-  --result-item-warning-border-color: #ff9f004d;
-  --result-item-question-color: #607d8b;
-  --result-item-question-border-color: #eceff1;
-  --result-item-empty-color: #9e9e9e;
-  --result-item-empty-border-color: #9e9e9e4d;
-}
-
-.var-result-item {
-  margin: 0 auto;
-  position: relative;
-  border-radius: 50%;
-  box-sizing: content-box;
-  border: 4px solid;
-
-  svg {
-    margin: 0 auto;
-    display: block;
-    width: 100%;
-    height: 100%;
-  }
-
-  &--info {
-    border-color: var(--result-item-info-border-color);
-
-    path {
-      fill: var(--result-item-info-color);
-    }
-  }
-
-  &--empty {
-    border-color: var(--result-item-empty-border-color);
-
-    path {
-      fill: var(--result-item-empty-color);
-    }
-  }
-
-  &--question {
-    border-color: var(--result-item-question-border-color);
-
-    path {
-      fill: var(--result-item-question-color);
-    }
-  }
-
-  &--error {
-    border-color: var(--result-item-error-border-color);
-
-    path {
-      fill: var(--result-item-error-color);
-    }
-  }
-
-  &--warning {
-    border-color: var(--result-item-warning-border-color);
-
-    path {
-      fill: var(--result-item-warning-color);
-    }
-  }
-
-  &--success {
-    position: relative;
-    border-radius: 50%;
-    box-sizing: content-box;
-    border-color: var(--result-item-success-color);
-    display: flex;
-    justify-content: center;
-    user-select: none;
-    zoom: normal;
-
-    &::before {
-      content: ' ';
-      top: 3.75%;
-      left: 2.5%;
-      width: 37.5%;
-      transform-origin: 100% 50%;
-      border-radius: 125% 0 0 125%;
-    }
-
-    &--cover-right {
-      top: 0;
-      display: block;
-      left: 37.5%;
-      width: 75%;
-      transform-origin: 0 50%;
-      border-radius: 0 125% 125% 0;
-      animation-name: rotate-circle;
-      animation-timing-function: ease-in;
-    }
-
-    &::before,
-    &--cover-right {
-      height: 125%;
-      position: absolute;
-      background: #fff;
-      transform: rotate(-45deg);
-    }
-
-    &--line {
-      //height: 6.25%;
-      height: 8.8%;
-      background-color: var(--result-item-success-color);
-      display: block;
-      border-radius: 2.5%;
-      position: absolute;
-      z-index: 10;
-    }
-
-    &--line-tip {
-      top: 57.5%;
-      left: 18.5%;
-      width: 31.25%;
-      transform: rotate(45deg);
-      animation-name: success-line-tip;
-    }
-
-    &--line-long {
-      top: 47.5%;
-      right: 10.4%;
-      width: 58.75%;
-      transform: rotate(-45deg);
-      animation-name: success-line-long;
-    }
-
-    @keyframes success-line-tip {
-      0% {
-        width: 0;
-        left: 1.25%;
-        top: 25%;
-      }
-      54% {
-        width: 0;
-        left: 1.25%;
-        top: 25%;
-      }
-      70% {
-        width: 62.5%;
-        left: -10%;
-        top: 46.25%;
-      }
-      84% {
-        width: 21.25%;
-        left: 26.25%;
-        top: 60%;
-      }
-      100% {
-        width: 31.25%;
-        left: 18.5%;
-        top: 57.25%;
-      }
-    }
-
-    @keyframes success-line-long {
-      0% {
-        width: 0;
-        right: 62%;
-        top: 72%;
-      }
-      65% {
-        width: 0;
-        right: 62%;
-        top: 72%;
-      }
-      84% {
-        width: 68.75%;
-        right: 0;
-        top: 42.5%;
-      }
-      100% {
-        width: 58.75%;
-        right: 10%;
-        top: 47.5%;
-      }
-    }
-
-    &--circle {
-      top: -5%;
-      left: -5%;
-      z-index: 10;
-      width: 100%;
-      height: 100%;
-      border-radius: 50%;
-      position: absolute;
-      box-sizing: content-box;
-      border: 4px solid var(--result-item-success-border-color);
-    }
-
-    &--fix {
-      top: 10%;
-      width: 6.25%;
-      left: 32.5%;
-      z-index: 1;
-      height: 106.25%;
-      position: absolute;
-      transform: rotate(-45deg);
-      background-color: #fff;
-    }
-
-    @keyframes rotate-circle {
-      0% {
-        transform: rotate(-45deg);
-      }
-      5% {
-        transform: rotate(-45deg);
-      }
-      12% {
-        transform: rotate(-405deg);
-      }
-      100% {
-        transform: rotate(-405deg);
-      }
-    }
-  }
-}
