+++ packages/varlet-ui/src/result/Info.vue
