+++ packages/varlet-ui/src/result/Warning.vue
